COL-1781, extend socket-io payload when /create elements
